reorganize StreamRef implementation
